--- google-beta/resource_active_directory_domain_generated_test.go
@@ -1,92 +1,0 @@
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
-package google
-
-import (
-	"fmt"
-	"strings"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-)
-
-func TestAccActiveDirectoryDomain_activeDirectoryDomainBasicExample(t *testing.T) {
-	t.Parallel()
-
-	context := map[string]interface{}{
-		"random_suffix": randString(t, 10),
-	}
-
-	vcrTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckActiveDirectoryDomainDestroyProducer(t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccActiveDirectoryDomain_activeDirectoryDomainBasicExample(context),
-			},
-			{
-				ResourceName:            "google_active_directory_domain.ad-domain",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"domain_name"},
-			},
-		},
-	})
-}
-
-func testAccActiveDirectoryDomain_activeDirectoryDomainBasicExample(context map[string]interface{}) string {
-	return Nprintf(`
-resource "google_active_directory_domain" "ad-domain" {
-  domain_name       = "tfgen%{random_suffix}.org.com"
-  locations         = ["us-central1"]
-  reserved_ip_range = "192.168.255.0/24" 
-}
-`, context)
-}
-
-func testAccCheckActiveDirectoryDomainDestroyProducer(t *testing.T) func(s *terraform.State) error {
-	return func(s *terraform.State) error {
-		for name, rs := range s.RootModule().Resources {
-			if rs.Type != "google_active_directory_domain" {
-				continue
-			}
-			if strings.HasPrefix(name, "data.") {
-				continue
-			}
-
-			config := googleProviderConfig(t)
-
-			url, err := replaceVarsForTest(config, rs, "{{ActiveDirectoryBasePath}}{{name}}")
-			if err != nil {
-				return err
-			}
-
-			billingProject := ""
-
-			if config.BillingProject != "" {
-				billingProject = config.BillingProject
-			}
-
-			_, err = sendRequest(config, "GET", billingProject, url, config.userAgent, nil)
-			if err == nil {
-				return fmt.Errorf("ActiveDirectoryDomain still exists at %s", url)
-			}
-		}
-
-		return nil
-	}
-}
