+++ google-beta/resource_active_directory_domain_update_test.go
@@ -2,12 +2,17 @@ package google
 
 import (
 	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 func TestAccActiveDirectoryDomain_update(t *testing.T) {
+	// skip the test until Active Directory setup issue got resolved
+	t.Skip()
+
 	t.Parallel()
 
 	domain := fmt.Sprintf("tf-test%s.org1.com", randString(t, 5))
@@ -77,4 +82,37 @@ func testAccADDomainUpdate(context map[string]interface{}) string {
 	}
 	`, context)
 
+}
+
+func testAccCheckActiveDirectoryDomainDestroyProducer(t *testing.T) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		for name, rs := range s.RootModule().Resources {
+			if rs.Type != "google_active_directory_domain" {
+				continue
+			}
+			if strings.HasPrefix(name, "data.") {
+				continue
+			}
+
+			config := googleProviderConfig(t)
+
+			url, err := replaceVarsForTest(config, rs, "{{ActiveDirectoryBasePath}}{{name}}")
+			if err != nil {
+				return err
+			}
+
+			billingProject := ""
+
+			if config.BillingProject != "" {
+				billingProject = config.BillingProject
+			}
+
+			_, err = sendRequest(config, "GET", billingProject, url, config.userAgent, nil)
+			if err == nil {
+				return fmt.Errorf("ActiveDirectoryDomain still exists at %s", url)
+			}
+		}
+
+		return nil
+	}
 }
